Match existing users also when creating accounts for unknown users is disabled
Moved the check for webserver_auth_create_user (which is FALSE by
default) from before checking if just user record in authmap table is
missing and matching an existing user depending on the value of
webserver_auth_match_existing (which is TRUE by default) after these
tasks right before creating the account for an unknown user depending on
the value of webserver_auth_create_user. This allows adding a record to
the authmap table for an existing user to authenticate using this module
without the need for enabling creating completely new accounts for
unknown users. The present state was inevitably resulting in the
administrator and all existing users being locked out of Drupal site
upon enabling the module, what was contrary to the intention of the
developers, as is clearly stated by the comment above the check for the
value of webserver_auth_match_existing.