GSSAPI authentication is now optional + return user login/logout to default state and re-enable login form at module's uninstallation
Created new configuration variable webserver_auth_optional_external_auth
that determines whether the external authentication will be mandatory
and the module will function as until now (default), or it will be
optional - users will be able to log in either using username and
password which they enter into the login form, or externally by clicking
the 'Log in externally' button in the User menu (that leads to the newly
created user/external_login menu path). Webserver authentication's
logout functionality is also moved to user/external_logout menu path,
but user has to manually create a link to use it. It is also necessary
to clear all caches and rebuild menus when enabling or disabling this
option - this is most easily done using the Devel module. Additionally,
uninstallation of the module will now also return user login and logout
functionality to its default state and re-enable the login form - this
will prevent locking all users out of the site in case of uninstallation
of the module.