Retrieve additional information about the user and update it at every login
The module now provides an option to retrieve the user's e-mail from
REMOTE_USER_EMAIL. This can be enabled by setting
webserver_auth_email_from_idp to TRUE; the user account modification
form will not allow manual modification of e-mail in that case. This
patch also adds webserver_auth_ext_groups_roles_map configuration
variable textarea - when this is nonempty, the module parses it (its
format has to be ext_group:role;ext_group:role...), checks if
REMOTE_USER_GROUP_* variables contain some group(s) that are covered by
this mapping, and if there are matches, it assigns the roles that are
mapped by this mapping to the user's external groups to the user. This
functionality is wrapped in two new functions:
_webserver_auth_update_remote_user_info(), which is called from
_webserver_auth_attempt_login(), and
_webserver_auth_update_remote_user_roles(), which is called from the
_webserver_auth_update_remote_user_info(). Finally, this patch also
contains few very minor code tweaks that are purely visual, without any
functional effect, intended to bring additions and modifications from
the previous patches into line with the module's coding conventions.